backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
openai==1.3.7
transformers==4.36.2
torch==2.1.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank-bm25==0.2.2
pymupdf==1.23.8
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
pytesseract==0.3.10
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
concurrent-futures==3.1.1